fix: loop and draw pool in experiment function
Fixed experiment function, for loop wasn't iterating because of the lack
of range method for the num_experiments. Also, fixed draw pool, the draw
method for the hat class changed the class variable contents, had to
create a copy of it before doing the draw so that the original sample
would not be affected by the draw.